refactor: rename HTMLSnapshotSource to ElementImageSource
Renames `HTMLSnapshotSource` to `ElementImageSource` to better align with the underlying `ElementImage` type from the HTML-in-Canvas proposal.

Key changes:
- Renamed class and all references throughout the codebase
- Moved shared types (`ElementImage`, `HTMLSourceCanvas`, `HTMLSourceResource`, `HTMLUploadableSource`) into new `HTMLSourceTypes.ts` file
- Simplified uploader logic by removing size-matching branches (always use 5-arg form)
- Updated element size calculation in `HTMLSource` to use `offsetWidth`/`offsetHeight` directly
- Made canvas ownership stricter: now throws if canvas cannot be determined
- Exported new `ElementImageSource` and `HTMLSourceTypes` from rendering index

The rename better reflects that this source wraps browser-provided element snapshots rather than "HTML" content specifically, matching the terminology from the experimental spec.
