+++ src/rendering/index.ts
@@ -217,9 +217,10 @@ export * from './renderers/shared/texture/sources/BufferImageSource';
 export * from './renderers/shared/texture/sources/CanvasSource';
 export * from './renderers/shared/texture/sources/CompressedSource';
 export * from './renderers/shared/texture/sources/CubeTextureSource';
+export * from './renderers/shared/texture/sources/ElementImageSource';
 export * from './renderers/shared/texture/sources/ExternalSource';
-export * from './renderers/shared/texture/sources/HTMLSnapshotSource';
 export * from './renderers/shared/texture/sources/HTMLSource';
+export * from './renderers/shared/texture/sources/HTMLSourceTypes';
 export * from './renderers/shared/texture/sources/ImageSource';
 export * from './renderers/shared/texture/sources/TextureSource';
 export * from './renderers/shared/texture/sources/VideoSource';
