+++ src/rendering/init.ts
@@ -4,7 +4,7 @@ import { ColorMask } from './mask/color/ColorMask';
 import { StencilMask } from './mask/stencil/StencilMask';
 import { BufferImageSource } from './renderers/shared/texture/sources/BufferImageSource';
 import { CanvasSource } from './renderers/shared/texture/sources/CanvasSource';
-import { HTMLSnapshotSource } from './renderers/shared/texture/sources/HTMLSnapshotSource';
+import { ElementImageSource } from './renderers/shared/texture/sources/ElementImageSource';
 import { HTMLSource } from './renderers/shared/texture/sources/HTMLSource';
 import { ImageSource } from './renderers/shared/texture/sources/ImageSource';
 import { VideoSource } from './renderers/shared/texture/sources/VideoSource';
@@ -20,6 +20,6 @@ extensions.add(
     CanvasSource,
     BufferImageSource,
     HTMLSource,
-    HTMLSnapshotSource,
+    ElementImageSource,
 );
